+++ app/src/main/java/com/lyddon/morgan/wikipedia/activities/MainActivity.kt
@@ -5,36 +5,46 @@ import android.os.Bundle
 import android.support.design.widget.BottomNavigationView
 import android.support.v7.app.AppCompatActivity
 import com.lyddon.morgan.wikipedia.R
+import com.lyddon.morgan.wikipedia.activities.fragments.ExploreFragment
+import com.lyddon.morgan.wikipedia.activities.fragments.FavoritesFragment
+import com.lyddon.morgan.wikipedia.activities.fragments.HistoryFragment
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
+    private val exploreFragment : ExploreFragment
+    private val favoritesFragment : FavoritesFragment
+    private val historyFragment : HistoryFragment
+
+    init {
+        exploreFragment = ExploreFragment()
+        favoritesFragment = FavoritesFragment()
+        historyFragment = HistoryFragment()
+    }
     private val mOnNavigationItemSelectedListener = BottomNavigationView.OnNavigationItemSelectedListener { item ->
-        when (item.itemId) {
-            R.id.navigation_home -> {
-                message.setText(R.string.title_home)
-                return@OnNavigationItemSelectedListener true
-            }
-            R.id.navigation_dashboard -> {
-                message.setText(R.string.title_dashboard)
-                return@OnNavigationItemSelectedListener true
-            }
-            R.id.navigation_notifications -> {
-                message.setText(R.string.title_notifications)
-                return@OnNavigationItemSelectedListener true
-            }
+
+        val transaction = supportFragmentManager.beginTransaction()
+        transaction.setCustomAnimations(android.R.anim.fade_in, android.R.anim.fade_out)
+
+        when(item.itemId){
+            R.id.navigation_explore -> transaction.replace(R.id.fragement_container, exploreFragment)
+            R.id.navigation_favorites -> transaction.replace(R.id.fragement_container, favoritesFragment)
+            R.id.navigation_history -> transaction.replace(R.id.fragement_container, historyFragment)
         }
-        false
+
+        transaction.commit()
+
+        true
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        message.setOnClickListener { _ ->
-            startActivity(Intent(this, ArticleDetailActivity::class.java))
-        }
-
         navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener)
+
+        val transaction = supportFragmentManager.beginTransaction()
+        transaction.add(R.id.fragement_container, exploreFragment)
+        transaction.commit()
     }
 }
