+++ app/src/main/java/com/lyddon/morgan/wikipedia/activities/fragments/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package com.lyddon.morgan.wikipedia.activities
+package com.lyddon.morgan.wikipedia.activities.fragments
 
 
 import android.os.Bundle
@@ -18,12 +18,12 @@ private const val ARG_PARAM2 = "param2"
  * A simple [Fragment] subclass.
  *
  */
-class ExploreFragment : Fragment() {
+class FavoritesFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_explore, container, false)
+        return inflater.inflate(R.layout.fragment_favorites, container, false)
     }
