+++ app/src/main/res/layout/fragment_explore.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".activities.ExploreFragment">
+    tools:context=".activities.fragments.ExploreFragment">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
