+++ app/src/main/res/layout/fragment_favorites.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.FavoritesFragment">
+    tools:context=".activities.fragments.FavoritesFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/favorites_article_recycler"
