+++ app/src/main/res/layout/fragment_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.HistoryFragment">
+    tools:context=".activities.fragments.HistoryFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/history_article_recycler"
