+++ app/src/main/res/menu/navigation.xml
@@ -2,18 +2,18 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:id="@+id/navigation_explore"
+        android:icon="@drawable/baseline_explore_white_24"
+        android:title="@string/title_explore" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:id="@+id/navigation_favorites"
+        android:icon="@drawable/baseline_favorite_white_24"
+        android:title="@string/title_favorites" />
 
     <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:id="@+id/navigation_history"
+        android:icon="@drawable/baseline_history_white_24"
+        android:title="@string/title_history" />
 
 </menu>
